[hw05-mongo] rename /api/course -> /api/courses
