+++ node-2023-11/hw04_les07-webServers/api-tests/framework/services/Courses.service.js
@@ -4,33 +4,33 @@ const { url } = require('../config');
 class CourseApi {
     async get() {
         const r = await supertest(`${url.localCoursesApi}`)
-            .get(`/api/course`)
+            .get(`/api/courses`)
             .set('Accept', 'application/json');
         return r;
     }
     async post(params) {
         const r = await supertest(`${url.localCoursesApi}`)
-            .post(`/api/course`)
+            .post(`/api/courses`)
             .set('Accept', 'application/json')
             .send(params);
         return r;
     }
     async getById(id) {
         const r = await supertest(`${url.localCoursesApi}`)
-            .get(`/api/course/${id}`)
+            .get(`/api/courses/${id}`)
             .set('Accept', 'application/json');
         return r;
     }
     async put(params) {
         const r = await supertest(`${url.localCoursesApi}`)
-            .put(`/api/course/${params.id}`)
+            .put(`/api/courses/${params.id}`)
             .set('Accept', 'application/json')
             .send(params.params);
         return r;
     }
     async delete(id) {
         const r = await supertest(`${url.localCoursesApi}`)
-            .delete(`/api/course/${id}`)
+            .delete(`/api/courses/${id}`)
             .set('Accept', 'application/json');
         return r;
     }
