+++ node-2023-11/hw04_les07-webServers/api-tests/framework/services/index.js
@@ -1,6 +1,6 @@
 const { UsersApi } = require('./Users.service');
 const { ResetApi } = require('./Reset.service');
-const { CourseApi } = require('./Course.service');
+const { CourseApi } = require('./Courses.service');
 
 const apiProvider = () => ({
     users: () => new UsersApi(),
